+++ packages/services/f-analytics/src/services/_tests/analytics.service.test.js
@@ -334,13 +334,12 @@ describe('Analytic Service ::', () => {
                 mockWindow({ winWidth, winHeight });
 
                 // Act
-                service.pushPageData({ pageName: 'jazz', groupName: 'man' });
+                service.pushPageData({ pageName: 'jazz' });
 
                 // Assert
                 expect(windowsPushSpy).toHaveBeenCalledWith(expect.objectContaining({
                     pageData: expect.objectContaining({
                         name: 'jazz',
-                        group: 'man',
                         orientation: orientationExpected,
                         display: displayExpected
                     })